+++ schema.dart
@@ -325,6 +325,10 @@ class AndroidTakeNativeScreenshotRequest {
   late String tag;
 }
 
+class IOSTakeNativeScreenshotRequest {
+  late String tag;
+}
+
 class IOSTakeCameraPhotoRequest {
   IOSSelector? shutterButtonSelector;
   IOSSelector? doneButtonSelector;
@@ -460,6 +464,9 @@ abstract class IosAutomator<IOSServer, DartClient> {
   void tapAt(IOSTapAtRequest request);
   void waitUntilVisible(IOSWaitUntilVisibleRequest request);
   void swipe(IOSSwipeRequest request);
+
+  // screenshots
+  void takeNativeScreenshot(IOSTakeNativeScreenshotRequest request);
 
   // notifications
   void closeHeadsUpNotification();